+++ chb/arm/opcodes/ARMLogicalShiftLeft.py
@@ -163,7 +163,7 @@ class ARMLogicalShiftLeft(ARMOpcode):
             bytestring=bytestring,
             annotations=annotations)
 
-        astree.add_reg_definition(iaddr, str(lhs), hl_rhs)
+        astree.add_reg_definition(iaddr, hl_lhs, hl_rhs)
         astree.add_instr_mapping(hl_assign, ll_assign)
         astree.add_instr_address(hl_assign, [iaddr])
         astree.add_expr_mapping(hl_rhs, ll_lsl_expr)