+++ chb/arm/opcodes/ARMStoreMultipleDecrementBefore.py
@@ -25,7 +25,7 @@
 # SOFTWARE.
 # ------------------------------------------------------------------------------
 
-from typing import cast, List, TYPE_CHECKING
+from typing import cast, Dict, List, Tuple, TYPE_CHECKING
 
 from chb.app.InstrXData import InstrXData
 
@@ -33,6 +33,7 @@ from chb.arm.ARMDictionaryRecord import armregistry
 from chb.arm.ARMOpcode import ARMOpcode, simplify_result
 from chb.arm.ARMOperand import ARMOperand
 
+from chb.ast.AbstractSyntaxTree import nooffset
 import chb.ast.ASTNode as AST
 from chb.astinterface.ASTInterface import ASTInterface
 
@@ -60,6 +61,21 @@ class ARMStoreMultipleDecrementBefore(ARMOpcode):
     args[2]: index of registers in arm dictionary
     args[3]: index of base memory address
     args[4]: thumb-wide
+
+    xdata format:vv[n]xxxx[2n]r[n+1]d[n+1][h[n+1]
+    vars[0]: base
+    vars[1..n]: lhs memory locations where values are stored
+    xprs[0]: base
+    xprs[1]: updated base (may be unchanged in case of no writeback)
+    xprs[2]: updated base (simplified)
+    xprs[3..n+2]: values of the registers being stored
+    xprs[n+3..2n+2]: values of the registers being stored (simplified)
+    rdefs[0]: reaching def base register
+    rdefs[1..n]: reaching defs registers being stored
+    defuse: use of base regster
+    defuse[1..n]: use of memory variables
+    defusehigh: use-hign of base register
+    defusehigh: use-high of memory variables
     """
 
     def __init__(
@@ -73,22 +89,29 @@ class ARMStoreMultipleDecrementBefore(ARMOpcode):
     def operands(self) -> List[ARMOperand]:
         return [self.armd.arm_operand(i) for i in self.args[1:-1]]
 
+    @property
+    def opargs(self) -> List[ARMOperand]:
+        return [self.armd.arm_operand(i) for i in self.args[1:-1]]
+
+    @property
+    def writeback(self) -> bool:
+        return self.args[0] == 1
+
     def is_store_instruction(self, xdata: InstrXData) -> bool:
         return True
 
     def annotation(self, xdata: InstrXData) -> str:
-        """xdata format: a:vxx .
-
-        vars[0..n-1]: lhs expressions
-        xprs[0..n-1]: rhs expressions
-        xprs[n]: initial value of base register
-        """
-
+        regcount = len(xdata.vars) - 1
+        wb = ""
+        if self.writeback:
+            wb = "; " + str(xdata.vars[0]) + " := " + str(xdata.xprs[2])
         return (
             "; ".join(
                 str(v)
                 + " := "
-                + str(x) for (v, x) in zip(xdata.vars, xdata.xprs[:-1])))
+                + str(x) for (v, x) in
+                zip(xdata.vars[1:], xdata.xprs[regcount+3:(2 * regcount)+3]))
+            + wb)
 
     def assembly_ast(
             self,
@@ -164,8 +187,8 @@ class ARMStoreMultipleDecrementBefore(ARMOpcode):
                             lhs = astree.mk_stack_variable_lval(
                                 offset, vtype=fieldtype)
                             fieldoffset = astree.mk_field_offset(field.fieldname, fieldtype)
-                            rhs0 = cast(AST.ASTLvalExpr, rhss[0])
                             rhslval = astree.mk_lval(rhs0.lval.lhost, fieldoffset)
+                            rhs0 = cast(AST.ASTLvalExpr, rhss[0])
                             rhs = astree.mk_lval_expr(rhslval)
                             instrs.append(astree.mk_assign(lhs, rhs, annotations=annotations))
                             offset += fieldtype.byte_size()
@@ -204,3 +227,238 @@ class ARMStoreMultipleDecrementBefore(ARMOpcode):
         '''
 
         return instrs
+
+    def ast_prov(
+            self,
+            astree: ASTInterface,
+            iaddr: str,
+            bytestring: str,
+            xdata: InstrXData) -> Tuple[
+                List[AST.ASTInstruction], List[AST.ASTInstruction]]:
+
+        regcount = len(xdata.reachingdefs) - 1
+        baselhs = xdata.vars[0]
+        memlhss = xdata.vars[1:]
+        baserhs = xdata.xprs[0]
+        baseresult = xdata.xprs[1]
+        baseresultr = xdata.xprs[2]
+        regrhss = xdata.xprs[3:3 + regcount]
+        rregrhss = xdata.xprs[3 + regcount:3 + (2 * regcount)]
+        baserdef = xdata.reachingdefs[0]
+        regrdefs = xdata.reachingdefs[1:]
+        baseuses = xdata.defuses[0]
+        memuses = xdata.defuses[1:]
+        baseuseshigh = xdata.defuseshigh[0]
+        memuseshigh = xdata.defuseshigh[1:]
+
+        annotations: List[str] = [iaddr, "STMDB"]
+
+        ll_wbackinstrs: List[AST.ASTInstruction] = []
+        hl_wbackinstrs: List[AST.ASTInstruction] = []
+
+        (baselval, _, _) = self.opargs[0].ast_lvalue(astree)
+        (baserval, _, _) = self.opargs[0].ast_rvalue(astree)
+
+        if self.writeback:
+            basedecr = astree.mk_integer_constant(4 * regcount)
+            ll_base_rhs = astree.mk_binary_op("minus", baserval, basedecr)
+            ll_base_assign = astree.mk_assign(baselval, ll_base_rhs, iaddr=iaddr)
+            ll_wbackinstrs.append(ll_base_assign)
+
+            hl_base_lhss = XU.xvariable_to_ast_lvals(baselhs, xdata, astree)
+            hl_base_rhss = XU.xxpr_to_ast_exprs(baseresultr, xdata, astree)
+            if len (hl_base_lhss) != 1 or len(hl_base_rhss) != 1:
+                raise UF.CHBError(
+                    "StoreMultipleDecrementBefore (STMDB): error in wback assign")
+            hl_base_lhs = hl_base_lhss[0]
+            hl_base_rhs = hl_base_rhss[0]
+            hl_base_assign = astree.mk_assign(hl_base_lhs, hl_base_rhs, iaddr=iaddr)
+            hl_wbackinstrs.append(hl_base_assign)
+
+            astree.add_instr_mapping(hl_base_assign, ll_base_assign)
+            astree.add_instr_address(hl_base_assign, [iaddr])
+            astree.add_expr_mapping(hl_base_rhs, ll_base_rhs)
+            astree.add_lval_mapping(hl_base_lhs, baselval)
+            astree.add_expr_reachingdefs(ll_base_rhs, [baserdef])
+            astree.add_lval_defuses(hl_base_lhs, baseuses)
+            astree.add_lval_defuses_high(hl_base_lhs, baseuseshigh)
+
+        def default() -> Tuple[List[AST.ASTInstruction], List[AST.ASTInstruction]]:
+            ll_instrs: List[AST.ASTInstruction] = []
+            hl_instrs: List[AST.ASTInstruction] = []
+            regsop = self.opargs[1]
+            registers = regsop.registers
+            base_decr = 4 * regcount
+            base_offset = base_decr
+            for (i, r) in enumerate(registers):
+                base_offset_c = astree.mk_integer_constant(base_offset)
+                addr = astree.mk_binary_op("minus", baserval, base_offset_c)
+                ll_lhs = astree.mk_memref_lval(addr)
+                ll_rhs = astree.mk_register_variable_expr(r)
+                ll_assign = astree.mk_assign(ll_lhs, ll_rhs, iaddr=iaddr)
+                ll_instrs.append(ll_assign)
+
+                lhs = memlhss[i]
+                rhs = rregrhss[i]
+                hl_lhss = XU.xvariable_to_ast_lvals(lhs, xdata, astree)
+                hl_rhss = XU.xxpr_to_ast_exprs(rhs, xdata, astree)
+                if len(hl_lhss) != 1 and len(hl_rhss) != 1:
+                    raise UF.CHBError(
+                        "StoreMultipleDecrementBefore (STMDB): error in assigns")
+                hl_lhs = hl_lhss[0]
+                hl_rhs = hl_rhss[0]
+                hl_assign = astree.mk_assign(hl_lhs, hl_rhs, iaddr=iaddr)
+                hl_instrs.append(hl_assign)
+
+                astree.add_instr_mapping(hl_assign, ll_assign)
+                astree.add_instr_address(hl_assign, [iaddr])
+                astree.add_expr_mapping(hl_rhs, ll_rhs)
+                astree.add_lval_mapping(hl_lhs, ll_lhs)
+                astree.add_expr_reachingdefs(ll_rhs, [regrdefs[i]])
+                astree.add_lval_defuses(hl_lhs, memuses[i])
+                astree.add_lval_defuses_high(hl_lhs, memuseshigh[i])
+
+                base_offset -= 4
+
+            return (hl_instrs + hl_wbackinstrs, ll_instrs + ll_wbackinstrs)
+
+        ll_instrs: List[AST.ASTInstruction] = []
+        hl_instrs: List[AST.ASTInstruction] = []
+
+        rhss = XU.xxpr_list_to_ast_exprs(rregrhss, xdata, astree)
+
+        if len(rhss) != 1:
+            astree.add_diagnostic(
+                "StoreMultipleDecrementBefore (STMDB): "
+                + "none or multiple values for rhs: "
+                + ", ".join(str(r) for r in rhss)
+                + " (use default)")
+            return default()
+
+        rhs = rhss[0]
+        rhstype = rhs.ctype(astree.ctyper)
+        if rhstype is None:
+            astree.add_diagnostic(
+                "StoreMultipleDecrementBefore (STMDB): "
+                + "no type found for "
+                + str(rhs)
+                + " (use default)")
+            return default()
+
+        if not rhstype.is_compound:
+            astree.add_diagnostic(
+                "StoreMultipleDecrementBefore (STMDB): "
+                + "type is not a struct: "
+                + str(rhs.ctype(astree.ctyper))
+                + " use default")
+            return default()
+
+        rhstype = cast(AST.ASTTypComp, rhs.ctype(astree.ctyper))
+        compinfo = astree.compinfo(rhstype.compkey)
+
+        if not memlhss[0].is_memory_variable:
+            astree.add_diagnostic(
+                "StoreMultipleDecrementBefore (STMDB): "
+                + "first variable is not a memory var: " + str(memlhss[0]))
+            return default()
+
+        xvar = cast ("VMemoryVariable", memlhss[0].denotation)
+        if not xvar.base.is_local_stack_frame:
+            astree.add_diagnostic(
+                "StoreMultipleDecrementBefore (STMDB): "
+                "variable is not a stack variable: "
+                + str(xvar))
+            return default()
+
+        startingoffset = xvar.offset.offsetvalue()
+        offset = startingoffset
+        rhs0 = cast(AST.ASTLvalExpr, rhss[0])
+        localvarassigns: Dict[int, AST.ASTAssign] = {}
+
+        def localvarassigns_at(offset: int) -> List[AST.ASTAssign]:
+            result: List[AST.ASTAssign] = []
+            for (o, a) in localvarassigns.items():
+                if o >= offset and o < (offset + 4):
+                    result.append(a)
+            return result
+
+        for field in compinfo.fieldinfos:
+            fieldtype = astree.resolve_type(field.fieldtype)
+
+            if fieldtype.is_scalar:
+                lhs = astree.mk_stack_variable_lval(offset, vtype=fieldtype)
+                fieldoffset = astree.mk_field_offset(
+                    field.fieldname, compinfo.compkey)
+                rhslval = astree.mk_lval(rhs0.lval.lhost, fieldoffset)
+                rhs = astree.mk_lval_expr(rhslval)
+                fieldassign = astree.mk_assign(lhs, rhs, iaddr=iaddr)
+                localvarassigns[offset] = fieldassign
+                offset += astree.type_size_in_bytes(fieldtype)
+
+            elif fieldtype.is_array:
+                arraytype = cast(AST.ASTTypArray, fieldtype)
+                if not arraytype.has_constant_size:
+                    astree.add_diagnostic(
+                        "StoreMultipleDecrementBefore (STMDB): "
+                        "array type does not have constant size")
+                    return default()
+
+                arraysize = arraytype.size_value()
+
+                elttype = astree.resolve_type(arraytype.tgttyp)
+                eltsize = astree.type_size_in_bytes(elttype)
+                for index in range(0, arraysize):
+                    lhs = astree.mk_stack_variable_lval(offset, vtype=elttype)
+                    indexoffset = astree.mk_scalar_index_offset(index)
+                    varoffset = astree.mk_field_offset(
+                        field.fieldname, compinfo.compkey, offset=indexoffset)
+                    rhslval = astree.mk_lval(rhs0.lval.lhost, varoffset)
+                    rhs = astree.mk_lval_expr(rhslval)
+                    eltassign = astree.mk_assign(lhs, rhs, iaddr=iaddr)
+                    localvarassigns[offset] = eltassign
+                    offset += eltsize
+            else:
+                fieldsize = astree.type_size_in_bytes(fieldtype)
+                lhs = astree.mk_stack_variable_lval(offset, vtype=fieldtype)
+                fieldoffset = astree.mk_field_offset(
+                    field.fieldname, compinfo.compkey)
+                rhslval = astree.mk_lval(rhs0.lval.lhost, fieldoffset)
+                rhs = astree.mk_lval_expr(rhslval)
+                fieldassign = astree.mk_assign(lhs, rhs, iaddr=iaddr)
+                localvarassigns[offset] = fieldassign
+
+                offset += fieldsize
+
+        ll_instructions: List[AST.ASTInstruction] = []
+        hl_instructions: List[AST.ASTInstruction] = []
+
+        regsop = self.opargs[1]
+        registers = regsop.registers
+        base_offset = startingoffset + (4 * regcount)
+        for (i, r) in enumerate(registers):
+            ll_offset = 4 * (4 - i)
+            realoffset = base_offset - ll_offset
+            hl_assigns = localvarassigns_at(realoffset)
+            ll_offset_c = astree.mk_integer_constant(ll_offset)
+            addr = astree.mk_binary_op("minus", baserval, ll_offset_c)
+            ll_lhs = astree.mk_memref_lval(addr)
+            ll_rhs = astree.mk_register_variable_expr(r)
+            ll_assign = astree.mk_assign(ll_lhs, ll_rhs, iaddr=iaddr)
+
+            ll_instructions.append(ll_assign)
+            hl_instructions.extend(hl_assigns)
+
+            for hl_asg in hl_assigns:
+                astree.add_local_vardefinition(iaddr, str(hl_asg.lhs), hl_asg.rhs)
+                astree.add_instr_mapping(hl_asg, ll_assign)
+                astree.add_instr_address(hl_asg, [iaddr])
+                astree.add_expr_mapping(hl_asg.rhs, ll_assign.rhs)
+                astree.add_lval_mapping(hl_asg.lhs, ll_assign.lhs)
+                astree.add_lval_defuses(hl_asg.lhs, memuses[i])
+                astree.add_lval_defuses_high(hl_asg.lhs, memuseshigh[i])
+
+            astree.add_expr_reachingdefs(ll_assign.rhs, [regrdefs[i]])
+
+        return (
+            hl_instructions + hl_wbackinstrs,
+            ll_instructions + ll_wbackinstrs)