+++ chb/ast/doc/CHANGELOG.md
@@ -6,6 +6,13 @@ The format is based on
 [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+### Version 0.1.0-2022-10-09
+
+- Added capability to add types to variables in the symboltable (replacing varinfos)
+- Added capability to resolve typedefs
+
+---
+
 ### Version 0.1.0-2022-10-06
 
 - Added created-by field to ast json file, with tool-name and version